Separated router and user authentication. Renamed login.js to user.js and now it only exports functions for the router to use.
